# Optimize the performance by passing MessageID implementations by pointers