--- a/pulsar/consumer_partition_test.go
+++ b/pulsar/consumer_partition_test.go
@@ -48,11 +48,11 @@ func TestSingleMessageIDNoAckTracker(t *testing.T) {
 	// ensure the tracker was set on the message id
 	messages := <-pc.queueCh
 	for _, m := range messages {
-		assert.Nil(t, m.ID().(trackingMessageID).tracker)
+		assert.Nil(t, m.ID().(*trackingMessageID).tracker)
 	}
 
 	// ack the message id
-	pc.AckID(messages[0].msgID.(trackingMessageID))
+	pc.AckID(messages[0].msgID.(*trackingMessageID))
 
 	select {
 	case <-eventsCh:
@@ -86,11 +86,12 @@ func TestBatchMessageIDNoAckTracker(t *testing.T) {
 	// ensure the tracker was set on the message id
 	messages := <-pc.queueCh
 	for _, m := range messages {
-		assert.Nil(t, m.ID().(trackingMessageID).tracker)
+		assert.Nil(t, m.ID().(*trackingMessageID).tracker)
 	}
 
 	// ack the message id
-	pc.AckID(messages[0].msgID.(trackingMessageID))
+	err := pc.AckID(messages[0].msgID.(*trackingMessageID))
+	assert.Nil(t, err)
 
 	select {
 	case <-eventsCh:
@@ -120,12 +121,13 @@ func TestBatchMessageIDWithAckTracker(t *testing.T) {
 	// ensure the tracker was set on the message id
 	messages := <-pc.queueCh
 	for _, m := range messages {
-		assert.NotNil(t, m.ID().(trackingMessageID).tracker)
+		assert.NotNil(t, m.ID().(*trackingMessageID).tracker)
 	}
 
 	// ack all message ids except the last one
 	for i := 0; i < 9; i++ {
-		pc.AckID(messages[i].msgID.(trackingMessageID))
+		err := pc.AckID(messages[i].msgID.(*trackingMessageID))
+		assert.Nil(t, err)
 	}
 
 	select {
@@ -135,7 +137,8 @@ func TestBatchMessageIDWithAckTracker(t *testing.T) {
 	}
 
 	// ack last message
-	pc.AckID(messages[9].msgID.(trackingMessageID))
+	err := pc.AckID(messages[9].msgID.(*trackingMessageID))
+	assert.Nil(t, err)
 
 	select {
 	case <-eventsCh: