@BewareMyPower 2023-02-28
Wait for https://github.com/apache/pulsar-client-go/pull/970 to be merged first.