@shibd 2023-03-01
> Also, why do you remove impl_message_bench_test.go?

This test has no assertions and should not be in unit tests.