@BewareMyPower 2023-03-01
> Do you have a bench result on this change? Since you declaim that it's a performance improvement.

I will paste a result soon. Though it might not be significant.

> Also, why do you remove impl_message_bench_test.go?

See my PR description.

> Remove the benchmark added in
https://github.com/apache/pulsar-client-go/pull/324. The result is obvious and this test is meaningless. I tried passing the trackingMessageID by pointer and the result reduced from 8.548 ns/op to 1.628 ns/op. It's obvious because a pointer is only 8 bytes while a trackingMessageID is 64 bytes. The overhead of accessing by pointers is far less than copying the extra bytes.